% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateRandomForest.R
\name{evaluate_rf_model}
\alias{evaluate_rf_model}
\title{Evaluate a Random Forest Model on Test Data}
\usage{
evaluate_rf_model(
  rf_model,
  test_texts,
  test_labels,
  tfidf_model,
  vectorizer,
  stopwords,
  verbose = TRUE
)
}
\arguments{
\item{rf_model}{A trained `ranger` model object.}

\item{test_texts}{A vector of raw test texts.}

\item{test_labels}{A factor vector of true labels.}

\item{tfidf_model}{The TF-IDF transformer used for training.}

\item{vectorizer}{The vectorizer used to build DTM.}

\item{stopwords}{A character vector of stopwords.}

\item{verbose}{Whether to print progress. Default TRUE.}
}
\value{
A list with test accuracy, test predictions, and aligned test data.
}
\description{
Evaluate a Random Forest Model on Test Data
}
