% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_ranger.R
\name{train_rf_model}
\alias{train_rf_model}
\title{Train a Random Forest Model with TF-IDF Features}
\usage{
train_rf_model(
  train_matrix,
  train_labels,
  ntree = 300,
  mtry = NULL,
  seed = 123,
  verbose = TRUE,
  train_df_cache_path = "train_df_cached.rds"
)
}
\arguments{
\item{train_matrix}{A sparse matrix (`dgCMatrix`) of training features.}

\item{train_labels}{A factor vector of training labels.}

\item{ntree}{Number of trees. Default 300.}

\item{mtry}{Variables to consider at each split. If NULL, auto-selected.}

\item{seed}{Random seed. Default 123.}

\item{verbose}{Whether to print progress. Default TRUE.}

\item{train_df_cache_path}{Path to cache the train data frame. Default "train_df_cached.rds".}
}
\value{
A trained `ranger` model object.
}
\description{
Train a Random Forest Model with TF-IDF Features
}
